requirements.txt
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
pydantic>=2.5.0
pyyaml>=6.0.1
colorama>=0.4.6
rich>=13.7.0
jinja2>=3.1.2
python-dotenv>=1.0.0
aiosignal>=1.3.1
asyncio>=3.4.3
